+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to the PF2E Visioner module will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2025-08-08
+
+### Changed
+- **Seek Template Player Logic**: Players will now be able to put their own seek templates, letting GMs open the results nicely
+- **Point out Player Logic**: Players are now able to point out themselves, letting the GM open the results nicely, added ping on the pointed out token
+
+
 ## [1.2.0] - 2025-08-08
 
 ### Added